pyproject.toml
[project]
name = "yt-transcribe"
version = "0.1.0"
description = "Transcribe and summarize video/audio content with AI - optimized for Apple Silicon Macs"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.0.0",
    "litellm>=1.0.0",
    "markdown>=3.5.0",
    "reportlab>=4.0.0",
    "python-telegram-bot>=21.0",
    "click>=8.1.0",
]

[project.scripts]
yt-transcribe = "yt_transcribe.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yt_transcribe"]
